requirements.txt
Flask>=3.0.0
pandas>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
pytest>=7.0.0
gunicorn>=21.0.0
